+++ trulens_eval/trulens_eval/app.py
@@ -17,18 +17,16 @@ from typing import (Any, Awaitable, Callable, ClassVar, Dict, Hashable,
 
 import pydantic
 
+from trulens_eval import app as mod_app
+from trulens_eval import feedback as mod_feedback
 from trulens_eval import schema as mod_schema
-from trulens_eval.feedback import Feedback
 from trulens_eval.instruments import Instrument
 from trulens_eval.instruments import WithInstrumentCallbacks
-from trulens_eval.schema import AppDefinition
+from trulens_eval.schema import app as mod_app_schema
 from trulens_eval.schema import Cost
-from trulens_eval.schema import FeedbackMode
-from trulens_eval.schema import FeedbackResult
+from trulens_eval.schema import feedback as mod_feedback_schema
 from trulens_eval.schema import Perf
-from trulens_eval.schema import Record
-from trulens_eval.schema import RecordAppCall
-from trulens_eval.schema import Select
+from trulens_eval.schema import record as mod_record
 from trulens_eval.utils import pyschema
 from trulens_eval.utils.asynchro import CallableMaybeAwaitable
 from trulens_eval.utils.asynchro import desync
@@ -351,11 +349,11 @@ class RecordingContext():
     - Combinations of the above.
     """
 
-    def __init__(self, app: App, record_metadata: JSON = None):
-        self.calls: List[RecordAppCall] = []
+    def __init__(self, app: mod_app.App, record_metadata: JSON = None):
+        self.calls: List[mod_record.RecordAppCall] = []
         """A record (in terms of its RecordAppCall) in process of being created."""
 
-        self.records: List[Record] = []
+        self.records: List[mod_record.Record] = []
         """Completed records."""
 
         self.lock: Lock = Lock()
@@ -373,7 +371,7 @@ class RecordingContext():
     def __iter__(self):
         return iter(self.records)
 
-    def get(self) -> Record:
+    def get(self) -> mod_record.Record:
         """
         Get the single record only if there was exactly one. Otherwise throw an error.
         """
@@ -389,7 +387,7 @@ class RecordingContext():
 
         return self.records[0]
 
-    def __getitem__(self, idx: int) -> Record:
+    def __getitem__(self, idx: int) -> mod_record.Record:
         return self.records[idx]
 
     def __len__(self):
@@ -403,7 +401,7 @@ class RecordingContext():
         return hash(self) == hash(other)
         # return id(self.app) == id(other.app) and id(self.records) == id(other.records)
 
-    def add_call(self, call: RecordAppCall):
+    def add_call(self, call: mod_record.RecordAppCall):
         """
         Add the given call to the currently tracked call list.
         """
@@ -412,8 +410,8 @@ class RecordingContext():
 
     def finish_record(
         self,
-        calls_to_record: Callable[[List[RecordAppCall], JSON], Record],
-        existing_record: Optional[Record] = None
+        calls_to_record: Callable[[List[mod_record.RecordAppCall], JSON], mod_record.Record],
+        existing_record: Optional[mod_record.Record] = None
     ):
         """
         Run the given function to build a record from the tracked calls and any
@@ -436,7 +434,7 @@ class RecordingContext():
         return record
 
 
-class App(AppDefinition, WithInstrumentCallbacks, Hashable):
+class App(mod_app_schema.AppDefinition, WithInstrumentCallbacks, Hashable):
     """Base app recorder type.
 
     Non-serialized fields here while the serialized ones are defined in
@@ -460,7 +458,7 @@ class App(AppDefinition, WithInstrumentCallbacks, Hashable):
         'arbitrary_types_allowed': True
     }
 
-    feedbacks: List[Feedback] = pydantic.Field(
+    feedbacks: List[mod_feedback.Feedback] = pydantic.Field(
         exclude=True, default_factory=list
     )
     """Feedback functions to evaluate on each record."""
@@ -507,7 +505,7 @@ class App(AppDefinition, WithInstrumentCallbacks, Hashable):
     """Mapping of instrumented methods (by id(.) of owner object and the
     function) to their path in this app."""
 
-    records_with_pending_feedback_results: Queue[Record] = \
+    records_with_pending_feedback_results: Queue[mod_record.Record] = \
         pydantic.Field(exclude=True, default_factory=lambda: Queue(maxsize=1024))
     """Records produced by this app which might have yet to finish
     feedback runs."""
@@ -535,7 +533,7 @@ class App(AppDefinition, WithInstrumentCallbacks, Hashable):
     def __init__(
         self,
         tru: Optional[Tru] = None,
-        feedbacks: Optional[Iterable[Feedback]] = None,
+        feedbacks: Optional[Iterable[mod_feedback.Feedback]] = None,
         **kwargs
     ):
         if feedbacks is not None:
@@ -562,7 +560,7 @@ class App(AppDefinition, WithInstrumentCallbacks, Hashable):
         else:
             pass
 
-        if self.feedback_mode == FeedbackMode.WITH_APP_THREAD:
+        if self.feedback_mode == mod_app_schema.FeedbackMode.WITH_APP_THREAD:
             self._start_manage_pending_feedback_results()
 
         self._tru_post_init()
@@ -666,13 +664,13 @@ class App(AppDefinition, WithInstrumentCallbacks, Hashable):
         """
 
         if self.tru is None:
-            if self.feedback_mode != FeedbackMode.NONE:
+            if self.feedback_mode != mod_app_schema.FeedbackMode.NONE:
                 from trulens_eval.tru import Tru
                 logger.debug("Creating default tru.")
                 self.tru = Tru()
 
         else:
-            if self.feedback_mode == FeedbackMode.NONE:
+            if self.feedback_mode == mod_app_schema.FeedbackMode.NONE:
                 logger.warning(
                     "`tru` is specified but `feedback_mode` is FeedbackMode.NONE. "
                     "No feedback evaluation and logging will occur."
@@ -683,7 +681,7 @@ class App(AppDefinition, WithInstrumentCallbacks, Hashable):
 
             self.db.insert_app(app=self)
 
-            if self.feedback_mode != FeedbackMode.NONE:
+            if self.feedback_mode != mod_app_schema.FeedbackMode.NONE:
                 logger.debug("Inserting feedback function definitions to db.")
 
                 for f in self.feedbacks:
@@ -695,7 +693,7 @@ class App(AppDefinition, WithInstrumentCallbacks, Hashable):
                     "Feedback logging requires `tru` to be specified."
                 )
 
-        if self.feedback_mode == FeedbackMode.DEFERRED:
+        if self.feedback_mode == mod_app_schema.FeedbackMode.DEFERRED:
             for f in self.feedbacks:
                 # Try to load each of the feedback implementations. Deferred
                 # mode will do this but we want to fail earlier at app
@@ -1118,12 +1116,12 @@ class App(AppDefinition, WithInstrumentCallbacks, Hashable):
         if record.feedback_and_future_results is None:
             return record
 
-        if self.feedback_mode == FeedbackMode.WITH_APP_THREAD:
+        if self.feedback_mode == mod_app_schema.FeedbackMode.WITH_APP_THREAD:
             # Add the record to ones with pending feedback.
 
             self.records_with_pending_feedback_results.put(record)
 
-        elif self.feedback_mode == FeedbackMode.WITH_APP:
+        elif self.feedback_mode == mod_app_schema.FeedbackMode.WITH_APP:
             # If in blocking mode ("WITH_APP"), wait for feedbacks to finished
             # evaluating before returning the record.
 
@@ -1285,7 +1283,7 @@ you use the `%s` wrapper to make sure `%s` does get instrumented. `%s` method
         )
 
     def _add_future_feedback(
-        self, future_or_result: Union[FeedbackResult, Future[FeedbackResult]]
+        self, future_or_result: Union[mod_feedback_schema.FeedbackResult, Future[mod_feedback_schema.FeedbackResult]]
     ) -> None:
         """
         Callback used to add feedback results to the database once they are
@@ -1303,9 +1301,9 @@ you use the `%s` wrapper to make sure `%s` does get instrumented. `%s` method
 
     def _handle_record(
         self,
-        record: Record,
-        feedback_mode: Optional[FeedbackMode] = None
-    ) -> Optional[List[Tuple[Feedback, Future[FeedbackResult]]]]:
+        record: mod_record.Record,
+        feedback_mode: Optional[mod_app_schema.FeedbackMode] = None
+    ) -> Optional[List[Tuple[mod_feedback.Feedback, Future[mod_feedback_schema.FeedbackResult]]]]:
         """
         Write out record-related info to database if set and schedule feedback
         functions to be evaluated. If feedback_mode is provided, will use that
@@ -1328,10 +1326,10 @@ you use the `%s` wrapper to make sure `%s` does get instrumented. `%s` method
             return []
 
         # Add empty (to run) feedback to db.
-        if feedback_mode == FeedbackMode.DEFERRED:
+        if feedback_mode == mod_app_schema.FeedbackMode.DEFERRED:
             for f in self.feedbacks:
                 self.db.insert_feedback(
-                    FeedbackResult(
+                    mod_feedback_schema.FeedbackResult(
                         name=f.name,
                         record_id=record_id,
                         feedback_definition_id=f.feedback_definition_id
@@ -1340,8 +1338,8 @@ you use the `%s` wrapper to make sure `%s` does get instrumented. `%s` method
 
             return None
 
-        elif feedback_mode in [FeedbackMode.WITH_APP,
-                               FeedbackMode.WITH_APP_THREAD]:
+        elif feedback_mode in [mod_app_schema.FeedbackMode.WITH_APP,
+                               mod_app_schema.FeedbackMode.WITH_APP_THREAD]:
 
             return self.tru._submit_feedback_functions(
                 record=record,
@@ -1350,7 +1348,7 @@ you use the `%s` wrapper to make sure `%s` does get instrumented. `%s` method
                 on_done=self._add_future_feedback
             )
 
-    def _handle_error(self, record: Record, error: Exception):
+    def _handle_error(self, record: mod_record.Record, error: Exception):
         if self.db is None:
             return
 
@@ -1381,7 +1379,7 @@ you use the `%s` wrapper to make sure `%s` does get instrumented. `%s` method
         main_error: str = "main_error are strings.",
         meta: Dict = {'metakey': 'meta are dicts'},
         tags: str = 'tags are strings'
-    ) -> Record:
+    ) -> mod_record.Record:
         """Create a dummy record with some of the expected structure without
         actually invoking the app.
 
@@ -1417,7 +1415,7 @@ you use the `%s` wrapper to make sure `%s` does get instrumented. `%s` method
                     else:
                         sample_args[p.name] = p.default
 
-                sample_call = RecordAppCall(
+                sample_call = mod_record.RecordAppCall(
                     stack=[
                         mod_schema.RecordAppCallMethod(
                             path=lens, method=method_serial
@@ -1431,7 +1429,7 @@ you use the `%s` wrapper to make sure `%s` does get instrumented. `%s` method
 
                 calls.append(sample_call)
 
-        return Record(
+        return mod_record.Record(
             app_id=self.app_id,
             calls=calls,
             cost=cost,
@@ -1470,7 +1468,7 @@ you use the `%s` wrapper to make sure `%s` does get instrumented. `%s` method
 
         return "\n".join(
             f"Object at 0x{obj:x}:\n\t" + "\n\t".
-            join(f"{m} with path {Select.App + path}"
+            join(f"{m} with path {mod_feedback_schema.Select.App + path}"
                  for m, path in p.items())
             for obj, p in self.instrumented_methods.items()
         )
