+++ trulens_eval/trulens_eval/schema/types.py
@@ -1,0 +1,23 @@
+from typing import Dict
+import typing_extensions
+
+# Identifier type aliases.
+
+RecordID: typing_extensions.TypeAlias = str
+"""Unique identifier for a record."""
+
+AppID: typing_extensions.TypeAlias = str
+"""Unique identifier for an app."""
+
+Tags: typing_extensions.TypeAlias = str
+"""Tags for an app or record."""
+
+Metadata: typing_extensions.TypeAlias = Dict
+"""Metadata for an app or record."""
+
+FeedbackDefinitionID: typing_extensions.TypeAlias = str
+"""Unique identifier for a feedback definition."""
+
+FeedbackResultID: typing_extensions.TypeAlias = str
+"""Unique identifier for a feedback result."""
+
